+++ projects/ngx-dt-table/src/lib/dt-filter-event.ts
@@ -1,0 +1,8 @@
+
+export interface DtFilterEvent {
+  filters: {
+    field: string,
+    comparison: string,
+    value: string | number | Date | {min: number | Date, max: number | Date} | any
+  }[];
+}